finished unsigned

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: danjose- <danjose-@student.42madrid.c      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/07 21:11:59 by danjose-          #+#    #+#             */
-/*   Updated: 2025/10/08 23:42:24 by danjose-         ###   ########.fr       */
+/*   Updated: 2025/10/11 01:54:41 by danjose-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ static	int	check_letter(va_list ap, char const letter)
 	if (letter == 'd' || letter == 'i')
 		i += ft_printnbr(va_arg(ap, int));
 	if (letter == 'u')
-		i += ft_printnbr_uns(va_arg(ap, long));
+		i += ft_printnbr_uns(va_arg(ap, long long));
 	return (i);
 }
